Use a local copy of main.sail and instrumentFns.sail.
Primarily this is to support CHERI versions of the config
variables set via __SetConfig.  Specifically, rip is now set to a
capablity containing the address rather than a null-derived value.

While here, permit setting all of the GPRs via config variables
instead of just rsp and rdi.  Also purge some bis that aren't very
general purpose.

Also while here, make instrumentFns.sail local to CHERI to
override the printState() function to log capabilities.

Ideally we would move printState() to a separate file so that the
CHERI model could override it, and similarly we might want to
restructure main.sail in the base x86 model so that used a hook to
set variables instead.